Change to list
Add aria to ensure the markup reads as grouped lists to AT yet maintain the Design System layout and styling without adding additional code. (This also potentially allows for more fine grain control of what is within grouped lists)

Explored using ol > li markup however decided against it due to the nested grouping that was needed to apply the Design System's rows and columns. 

This page is somewhat un-usual, in that this is list of images.  For example to convert to a list and use the grid system -  markup similar to this is needed which is invalid:

ol (21st century) > div (rows wrapper) > li (column)